requirements.txt
amqp==2.4.2
beautifulsoup4==4.7.1
billiard==3.6.0.0
bs4==0.0.1
celery==4.3.0
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-filters==1.0.0.dev0
kombu==4.5.0
lxml==4.3.3
pkg-resources==0.0.0
pytz==2019.1
soupsieve==1.9.1
sqlparse==0.3.0
vine==1.3.0